+++ integration-tests/src/test/kotlin/com/adobe/testing/s3mock/its/GetPutDeleteObjectV1IT.kt
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2017-2022 Adobe.
+ *  Copyright 2017-2023 Adobe.
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -151,6 +151,7 @@ internal class GetPutDeleteObjectV1IT : S3TestBase() {
     assertThat(objectMetadata.userMetadata)
       .`as`("User metadata should be identical!")
       .isEqualTo(metadata.userMetadata)
+    assertThat(objectMetadata.sseAwsKmsKeyId).isEqualTo(TEST_ENC_KEY_ID)
   }
 
   /**